requirements.txt
tqdm
pandas
matplotlib
gluoncv
gluoncv2
mxnet-cu101mkl
numpy
scikit-learn
